+++ api/database.json
@@ -86,7 +86,7 @@
       "name": "Testing",
       "completionDate": "2020-01-20",
       "id": 5,
-      "isCompleted": false
+      "isCompleted": true
     }
   ],
   "messages": [
